:recycle: [Refactor] CalendarSection 분리 #easy
