+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Box, Flex } from '@chakra-ui/react';
 
 import { AlertSection } from './components/AlertSection';
-import { CalendarSection } from './components/CalendarSection';
+import { Calendar } from './components/calendar/Calendar';
 import { EventForm } from './components/eventForm/EventForm.tsx';
 import { EventListSection } from './components/EventListSection';
 
@@ -10,7 +10,7 @@ function App() {
     <Box w="full" h="100vh" m="auto" p={5}>
       <Flex gap={6} h="full">
         <EventForm />
-        <CalendarSection />
+        <Calendar />
         <EventListSection />
       </Flex>
       <AlertSection />
